pyproject.toml
[project]
name = "project-agents"
version = "0.1.0"
description = "AI-агент для управления бизнес-проектами через Telegram"
requires-python = ">=3.12"

dependencies = [
    "anthropic>=0.52",
    "mcp>=1.9",
    "aiogram>=3.25",
    "aiosqlite>=0.21",
    "pyyaml>=6.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "apscheduler>=3.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
